Fit 1.5: HomeViewModel & HomeView Refactor
