+++ FitnessApp/Home/Models/Activity.swift
@@ -1,0 +1,18 @@
+//
+//  Activity.swift
+//  FitnessApp
+//
+//  Created by Thanveer ahmed Shaik on 12/02/25.
+//
+
+import Foundation
+import SwiftUI
+
+struct Activity{
+    let id:Int
+    let title:String
+    let subtitle:String
+    let image:String
+    let tintColor: Color
+    let amount: String
+}
