+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,19 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Thanveer ahmed Shaik on 12/02/25.
+//
+import SwiftUI
+import Foundation
+
+
+struct Workout{
+    let id: Int
+    let title: String
+    let image: String
+    let tintcolor:Color
+    let duration: String
+    let date: String
+    let calories: String
+}
