+++ FitnessApp/Home/Views/ActivityCard.swift
@@ -7,19 +7,14 @@
 
 import SwiftUI
 
-//Creating a model to make the activity card more dynamic
 
-struct Activity{
-    let id:Int
-    let title:String
-    let subtitle:String
-    let image:String
-    let tintColor: Color
-    let amount: String
-}
+
+
 
 struct ActivityCard: View {
     @State var activity:Activity
+    
+    
     var body: some View {
         ZStack{
             Color(uiColor: .systemGray6)
