+++ FitnessApp/Home/Views/WorkoutCard.swift
@@ -6,19 +6,13 @@
 //
 
 import SwiftUI
-struct Workout{
-    let id: Int
-    let title: String
-    let image: String
-    let tintcolor:Color
-    let duration: String
-    let date: String
-    let calories: String
-    
-}
+
+
 struct WorkoutCard: View {
     @State var workout: Workout
     var body: some View {
+        
+        
         HStack{
             Image(systemName: workout.image)
                 .resizable()
